requirements.txt
# ===== Core Dependencies =====
Django==4.2.25
djangorestframework==3.15.2
django-cors-headers==4.3.1

# ===== Database - MongoDB =====
djongo>=1.3.6
pymongo>=4.0.0
dnspython>=2.2.0

# ===== Production Server =====
gunicorn==22.0.0

# ===== Security & Authentication =====
djangorestframework-simplejwt>=5.3.0

# ===== Monitoring & Logging (Optional) =====
# sentry-sdk>=1.40.0

# ===== Development & Testing =====
# pytest>=7.4.0
# pytest-django>=4.5.0
# coverage>=7.3.0
# factory-boy>=3.3.0

# ===== Code Quality =====
# black>=23.12.0
# flake8>=6.1.0
# isort>=5.13.0

# ===== Azure Storage (Optional) =====
# azure-storage-blob>=12.19.0
